requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
cffi==1.16.0
cryptography
Django==4.2.9
paramiko==3.4.0
ping3==4.0.4
pycparser==2.21
PyNaCl==1.5.0
sqlparse==0.4.4
tzdata==2023.4
requests
python-decouple
pillow
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter 
drf-yasg[validation]
python-decouple
